LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-visited-cssom-expected.txt


FAIL :link as scoped selector assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 255, 255)"
PASS :visited as scoped selector
PASS :not(:link) as scoped selector
FAIL :not(:visited) as scoped selector assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 255, 255)"
FAIL :link as scoping root assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 255, 255)"
PASS :visited as scoping root
FAIL :not(:visited) as scoping root assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 255, 255)"
PASS :not(:link) as scoping root
FAIL :link as scoping root, :scope assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 255, 255)"
PASS :visited as scoping root, :scope
FAIL :not(:visited) as scoping root, :scope assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 255, 255)"
PASS :not(:link) as scoping root, :scope
PASS :link as scoping limit
FAIL :visited as scoping limit assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 255, 255)"
FAIL :not(:link) as scoping limit assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 255, 255)"
PASS :not(:visited) as scoping limit
